requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
numpy
pillow
opencv-python-headless
matplotlib
